Comments added in code.
Comments were added to be ready for submission. Wish me luck.

+++ Chapter-1-Introduction-to-C++/Exercises/2_DivisionFix/DivisionFix/DivisionFix.cpp
@@ -2,8 +2,9 @@
 using namespace std;
 
 int main() {
+    // A simple fix is provided by changing the dividend into a double, so we get the decimals in the quotient.
     double numberOne = 50;
-    double numberTwo = 7;
+    int numberTwo = 7;
     cout << numberOne / numberTwo << endl;
 
     return 0;
